+++ pkg/clickhouse/structs.go
@@ -19,7 +19,7 @@ type Table struct {
 	Name     string `ch:"name"`
 	Engine   string `ch:"engine"`
 	// fields depends on `clickhouse-server` version
-	DataPath         string   `ch:"data_path"` // For legacy support
+	DataPath         string   `ch:"data_path"`
 	DataPaths        []string `ch:"data_paths"`
 	UUID             string   `ch:"uuid"`
 	CreateTableQuery string   `ch:"create_table_query"`